Add shared metadata loader
